chore(ui): added update password section to user profile
